ultimas correções
