+++ Entrega/exerc04.js
@@ -1,19 +1,28 @@
-//4. Crie uma função que receba um número e devolva uma soma progressiva. Por exemplo, recebendo o número 5, a função deve retornar 15, ou seja, o resultado de 1 + 2 + 3 + 4 + 5. 
-
 const read = require('readline-sync')
 
-let num = parseFloat(read.question("Digite um numero: "))
+const parametro1 = Number(read.question("Digite o primeiro parametro: "));
+const parametro2 = Number(read.question("Digite o segundo parametro: "));
+const parametro3 = Number(read.question("Digite o terceiro parametro: "));
 
+function retornaParametros(parametro1, parametro2, parametro3){
+    if(parametro1 && parametro2 && parametro3){
+        console.log (`${(parametro1 + parametro2) / parametro3}`)
 
+    } else if(!parametro1 && !parametro2 && !parametro3){
+        console.log ("Não recebeu parâmetro.")
+        
+    } else if (!parametro2 && !parametro3) {
+        console.log (parametro1)
 
-function somarNumero(){
-    let resultado = 1
-    for(let cont = num; cont > 1; cont--){
-        resultado += cont
+    } else if (!parametro3 && !parametro1) {
+        console.log (parametro2)
+
+    }else if (!parametro2 && !parametro1) {
+        console.log (parametro3)
+
+    }else {
+        console.log(parametro1 + parametro2 + parametro3)
     }
-     return resultado;
 }
 
-console.log(somarNumero(num))
-
-
+retornaParametros(parametro1, parametro2, parametro3)
