+++ Entrega/tempCodeRunnerFile.js
@@ -1,0 +1,1 @@
+((preco * desconto ) / 100)
